requirements.txt
gradio
tensorflow
psycopg2-binary
Pillow
gdown
matplotlib
